+++ spps-shared/src/test/java/de/elomagic/spps/shared/SimpleCryptCommandToolTest.java
@@ -66,7 +66,10 @@ class SimpleCryptCommandToolTest {
             p.load(reader);
         }
 
-        Assertions.assertEquals(2, p.keySet().size());
+        Assertions.assertEquals(3, p.keySet().size());
+        Assertions.assertTrue(p.keySet().contains("created"));
+        Assertions.assertTrue(p.keySet().contains("key"));
+        Assertions.assertTrue(p.keySet().contains("relocation"));
 
         // TODO Test Print
     }