+++ spps-shared/src/test/java/de/elomagic/spps/shared/SimpleCryptTest.java
@@ -52,7 +52,10 @@ class SimpleCryptTest {
             p.load(reader);
         }
 
-        Assertions.assertEquals(2, p.keySet().size());
+        Assertions.assertEquals(3, p.keySet().size());
+        Assertions.assertTrue(p.keySet().contains("created"));
+        Assertions.assertTrue(p.keySet().contains("key"));
+        Assertions.assertTrue(p.keySet().contains("relocation"));
     }
 
     @Test